requirements.txt
click==6.6
imageio==1.5
numpy==1.11.0
Pillow==3.2.0
PyTweening==1.0.3
selenium==2.53.2
